Name strlen of dim/var/attr is expected < NC_MAX_INT

+++ examples/C/global_attributes.c
@@ -123,7 +123,7 @@ int main(int argc, char** argv)
     sprintf(str_att, "Mon Aug 13 21:27:48 2018");
 
     /* make sure the time string are consistent among all processes */
-    MPI_Bcast(str_att, strlen(str_att), MPI_CHAR, 0, MPI_COMM_WORLD);
+    MPI_Bcast(str_att, (int)strlen(str_att), MPI_CHAR, 0, MPI_COMM_WORLD);
 
     err = ncmpi_put_att_text(ncid, NC_GLOBAL, "history", strlen(str_att),
                              &str_att[0]); ERR
